Reorganize the basic_federated Zephyr test
