+++ examples/zephyr/basic_federated/federated_receiver/prj.conf
@@ -2,16 +2,13 @@ CONFIG_NEWLIB_LIBC=y
 CONFIG_NETWORKING=y
 CONFIG_NET_TCP=y
 CONFIG_NET_IPV4=y
-CONFIG_NET_IPV6=y
 CONFIG_NET_SOCKETS=y
 CONFIG_POSIX_API=y
-
-CONFIG_TIMESLICING=y
-
-CONFIG_TEST_RANDOM_GENERATOR=y
-
+CONFIG_ASSERT=y
+CONFIG_STACK_CANARIES=y
+CONFIG_MAIN_STACK_SIZE=4096
 # Network address config
+
 CONFIG_NET_CONFIG_SETTINGS=y
 CONFIG_NET_CONFIG_NEED_IPV4=y
-CONFIG_NET_CONFIG_MY_IPV4_ADDR="192.0.2.1"
-CONFIG_NET_CONFIG_PEER_IPV4_ADDR="192.0.2.2"
+CONFIG_NET_CONFIG_MY_IPV4_ADDR="192.168.1.101"
