+++ examples/zephyr/basic_federated/federated_sender/CMakeLists.txt
@@ -5,7 +5,7 @@ set(BUILD_EXAMPLES OFF CAHCHE BOOL)
 find_package(Zephyr REQUIRED HINTS $ENV{ZEPHYR_BASE})
 project(reactor-uc-zephyr)
 
-add_subdirectory(../../../ reactor-uc)
+add_subdirectory(../../../../ reactor-uc)
 
-target_sources(app PRIVATE src/testing_fed_conn.c)
+target_sources(app PRIVATE src/sender.c)
 target_link_libraries(app PRIVATE reactor-uc)
