+++ examples/zephyr/basic_federated/federated_sender/prj.conf
